+++ db/migrate/20180830054139_add_conversion_speed_to_character.rb
@@ -1,0 +1,5 @@
+class AddConversionSpeedToCharacter < ActiveRecord::Migration[5.2]
+  def change
+    add_column :characters, :conversion_speed, :integer, default: 0
+  end
+end