+++ db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 2018_08_29_153007) do
+ActiveRecord::Schema.define(version: 2018_08_30_054139) do
 
   # These are extensions that must be enabled in order to support this database
   enable_extension "plpgsql"
@@ -52,6 +52,7 @@ ActiveRecord::Schema.define(version: 2018_08_29_153007) do
     t.string "lifestyle_override"
     t.string "character_sheet_url"
     t.integer "conversion_type", default: 0
+    t.integer "conversion_speed", default: 0
     t.index ["user_id"], name: "index_characters_on_user_id"
   end